[CHORE] add extension for preview, set enum size
